Move dataclasses to schemas module
