--- cache/news_report_dataclass.py
@@ -1,11 +1,0 @@
-from dataclasses import dataclass
-from typing import List
-
-from cache.news_dataclass import NewsReportData
-
-
-@dataclass
-class NewsReport:
-    date: str
-    news_items: List[NewsReportData]
-    created_at: str
