+++ cache/schemas.py
@@ -38,3 +38,20 @@ class NewSignupsResponse:
     value_daily: int
     value_weekly: int
     diff: Optional[float]
+
+@dataclass
+class NewsReportData:
+    id: str
+    headline: str
+    summary: str
+    url: str
+    peak_interest: int
+    current_interest: int
+    news_section: str
+    thumbnail: str | None
+
+@dataclass
+class NewsReport:
+    date: str
+    news_items: List[NewsReportData]
+    created_at: str
